Undo Fetcher model for now... we can't have outstanding cursors -- they block other transactions (who knew!?!).
Put in place limit params to methods that return fetchers so that they
could potentially ask for X at a time to limit resource use.

We need to figure out if this actually makes a difference anyway...

Then I need to decide where to implement the limits -- on the backend
side (by making a more sophisticated Fetcher class) or on the
front-end/controller side (by making the front-end always ask for a
specific number of e.g. recipes when asking for a list).