Massive overhaul of keyEditor to work with new world order.
We also allow editing of amounts, units and items, since often key
errors are parsing errors that have mixed up other elements of
ingredients as well.

There is still work to be done... (namely, adding fields to do en
masse changes of items, units and amounts via fields (as opposed to
via pointing and clicking on the tree itself).