Some changes made necessary by our recent work on Undo.
In particular, we need a way to ask if two RowObjects refer to the
same Row in the database (and we can't use the built in __eq__ because
that foobars hashability of RowObjects which is a bummer). We create
the row_equal method. In the case of sql (our only current
implementation), we use the rowid field for this as a fallback -- to
facilitate row equivalence even when there is no unique ID for a
table.