Replace compressed jQuery v1.4.2 by uncompressed v1.11.0.
It was pointed out by Debian bug #744707 [1] that shipping only the
compressed version violates the DFSG (Debian Free Software Guidelines).
As the web plugin isn't really relying on a lot of jQuery based code to
execute quickly, I think using the uncompressed version is just as fine.

Unfortunately, I'm currently not able to check if the updated jQuery breaks
the web plugin as I'm unable to run it due to a number of errors.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=744707