Add a possibility to tell ExporterMultirec not to open any files.
This is necessary to use external libraries which process the files by its
own and finally write the files to a given location. This makes the process
of implementing such a plugin more easy.

This also fixes a possible issue if self.out is not a string, then the
program may crash because closing a string does not work.