Add a plugin to support epub export of recipes.
Now I can read the recipes on my tablet :) The plugin uses the ebooklib
available at pip which makes creation of a epub very easy. In the futuere
a mobie export is planned within this library too.

For the moment this is a basic implementation, in the future we should
implement some more things for e.g. to set the language or a book title,
cover page, ordering, categories, etc. but this is the first step.

refs #57