Add util package to setup.py file
Installing from source and running gourmet from the command line caused
an error in the imports for gglobals.py.  Specifically, from 'util
import windows' threw an exception because the util package was not
installed.  This commit adds 'gourmet.util' to the packages list of the
setup() function in setup.py, which installs the corresponding sub-
package.