Fix notification message when import cancelled
When a user begins importing a website and cancels the import in the
interactive importer window the message "Recipe successfully imported"
is displayed.  This is untrue; the message should tell the user that
nothing was imported.  This pull request changes the displayed message
to "Import Unsuccessful" when an import is cancelled.

The behavior noted above is caused by a callback function, thread_done
in the threadManager.py module, assuming the import is successful.
When a user closes the interactive importer window this method hides
the previous "Downloading ..." message and updates it with "Recipe
successfully imported."  To avoid this assumption the thread_done
method is now only responsible for hiding the previous message.  A new
method, importer_thread_done, handles determining if 0, 1 or more
recipes were added and displays an appropriate message.  Making this a
separate function allows the importer (e.g. FoodNetworkParser) to be
connected and passed to the callback. The importer contains information
on how many recipes are to be added, which is not present in the
previously used importManager object.

Note:  The export and deletion code was also changed to call a new
notification display function, but no behavior was modified.

Fixes #832