Catch error when enabling nutrition plugin
If a user enables, disables and then enables the nutrition plugin, an
InvalidRequestError is thrown by sqlalchemy because the tables the
plugin creates during activation have already been constructed.  The
error only occurs when a user enables the plugin twice in the same
session.  This patch wraps table creation in a try except block since
the plugin still runs as expected.