testing to move to HTML5
I'm not a programmer, nor I'm abl eto compile so theese mods are not tested. 
Please read mi comments, they start as # AR...
These are my proposal for a more semantic (html5) output:
NAV (index button)
ARTICLE (the full recipe)
    HEADER the title
    FIGURE the image
    ASIDE
        TABLE of header elements
    SECTION
         TABLE of ingredients
    SECTION instructions
    SECTION notes
First and most important - the title is an H1 elment, not a P
moving DIVs to Html5 semantic tags
adding an "index" button that shall disappear in a one-page
head formatted as table - I wish to do it also for the ingredients but I wasn't able, sorry.