Sort recipes by main (first) category into sections
This allows to sensibly group multiple recipes in a book, and to
enable a structured table-of-content.

The first category, in the absence of a better mechanism, is treated
specially to primarily group recipes.  It is the first sorting
criteria and primary ToC entries are generated per first category.
The recipe name remains the secondary sorting criterion.

Fixes #57.

Signed-off-by: Martin Pohlack <martinp@gmx.de>