Sort recipes by main (first) category into sections
This allows to sensible group multiple recipes in a book, and enables
a structured table-of-content.

The first category, in the absence of another mechanism, is treated
specially to primarily group recipes.  It is the first sorting
criteria and primary section entries are generated per first
category.  Secondary sorting criteria is still the recipe name.

Signed-off-by: Martin Pohlack <martinp@gmx.de>