Save search actions toggle status
search as you type and regex search ToggleActions are now also saved in the prefs files for usage accross sessions. This makes the ToggleActions consistent with the widgets found in the settings menu. Previously, they were not saved, resulting in counter-intuitive application start-up where some actions were activated, but not shown in the GUI as in use.