Simpify installing a development environment
A development install can now be installed in one step using
'development.in'. This installs gourmet in "editable" mode as well as
additional development dependencies.

'development.txt' is renamed to 'development.in' to support pinning of
dependencies (using, e.g., pip-tools). Unfortunately, this only works
for local installs since gourmet is a "local" package and results in a
non-portable, absolute file path being generated in 'development.txt'.
Consequently, 'development.txt' is added to .gitignore, and build
environments can just use 'development.in' directly.

wheel is removed as a development dependency since it must already be
installed in order for pip to build wheels when installing dependencies.