Fix creating/updating PO templates
A missing space in the command passed to os.system meant this was
quietly failing since stderr was being surpressed. Unfortunately,
intltool-update seems to send all output to stderr and have no option to
quiet it, so there doesn't seem to be a way to make this quieter.
However, noisiness seems preferable to suppressing errors, and with
update_i18n now being run manually only when needed to update message
templates, this shouldn't be polluting any build logs.